Fixed a bug that had been introduced with the cycles: beacuse the number of seconds since the start of the epoch requires a certain precision in the floating point values, storing the time of death or time to depart as a float was not sufficient any more. This basically hints at a deeper problem with storing absolute times. Maybe everything should be converted to relative times.

+++ source/goals.h
@@ -251,7 +251,7 @@ public:
 	virtual void Process(void);
 private:
 	float _SecondsToWait;
-	float _TimeToLeave;
+	double _TimeToLeave;
 };
 
 #endif
